# swagger.yaml
swagger: "2.0"
info:
  description: "This is G2i Node Test"
  version: "1.0.0"
  title: "Acronym"
tags:
- name: acronyms
  description: acronyms API
host: "g2iacronymnodetest.herokuapp.com"
paths:
# [GET] acronyms
  /acronym:
    get:
      tags:
      - acronyms
      summary: Find All Acronyms
      parameters:
        - in: query
          name: from
          type: integer
        - in: query
          name: limit
          type: integer
        - in: query
          name: search
          type: string
      responses:
        200:
          description: 'OK'
        500:
          description: 'Server Error'
# [POST] acronyms
    post:
      tags:
      - acronyms
      summary: Add Acronym
      parameters:
      - name: body
        in: body
        description: acronym Data
        required: true
        schema:
          $ref: '#/definitions/createAcronym'
      responses:
        201:
          description: 'Created'
        400:
          description: 'Bad Request'
        409:
          description: 'Conflict'
        500:
          description: 'Server Error'
# [PUT] acronyms/name
  /acronym/{nowAcronym}:
    put:
      tags:
      - acronyms
      summary: Update Acronym By Name
      parameters:
      - name: Authorization
        in: header
        require: true
        type: string
      - in: path
        name: nowAcronym
        description: acronym Name
        required: true
        type: string
      - name: body
        in: body
        description: acronym Data
        required: true
        schema:
          $ref: '#/definitions/updateAcronym'
      responses:
        200:
          description: 'OK'
        400:
          description: 'Bad Request'
        409:
          description: 'Conflict'
        500:
          description: 'Server Error'
# [DELETE] acronyms/acronyms
    delete:
      tags:
      - acronyms
      summary: Delete Acronym By Name
      parameters:
      - name: nowAcronym
        in: path
        description: acronym Name
        required: true
        type: string
      - name: Authorization
        in: header
        require: true
        type: string
      responses:
        200:
          description: 'OK'
        409:
          description: 'Conflict'
        500:
          description: 'Server Error'
# definitions
definitions:
  createAcronym:
    type: object
    required:
        - acronym
        - description
    properties:
      acronym:
        type: string
        description: acronym Name
      description:
        type: string
        description: acronym description
  updateAcronym:
    type: object
    required:
        - newAcronym
        - newDescription
    properties:
      newAcronym:
        type: string
        description: acronym Name
      newDescription:
        type: string
        description: acronym description
schemes:
 - http
 - https